woops fix migration

+++ db/post_migrate/20260330161714_drop_pinned_post_number_from_nested_topics.rb
@@ -2,10 +2,14 @@
 
 class DropPinnedPostNumberFromNestedTopics < ActiveRecord::Migration[8.0]
   def up
-    remove_column :nested_topics, :pinned_post_number
+    if column_exists?(:nested_topics, :pinned_post_number)
+      remove_column :nested_topics, :pinned_post_number
+    end
   end
 
   def down
-    add_column :nested_topics, :pinned_post_number, :integer
+    unless column_exists?(:nested_topics, :pinned_post_number)
+      add_column :nested_topics, :pinned_post_number, :integer
+    end
   end
 end
